backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "python-dotenv>=1.0.0",
    "supabase>=2.3.0",
    "langchain>=0.1.0",
    "langgraph>=0.0.10",
    "langchain-ibm>=0.1.0",
    "fastembed>=0.2.0",
    "neo4j>=5.14.0",
    "firecrawl-py>=0.0.1",
    "sentence-transformers>=2.3.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "requests>=2.31.0",
    "sqlalchemy>=2.0.0",
    "pgvector>=0.2.0",
    "python-multipart>=0.0.9",
    "email_validator>=2.0.0",
]
